requirements.txt
Django==2.2
djangorestframework==3.12.2
#pytz==2021.1
#sqlparse==0.4.1
